feat : market view product success
장터 10개 품목 조회하는 부분 완료했습니다.

+++ finfarm-backend/src/main/java/com/moneygang/finfarm/domain/market/entity/Seed.java
@@ -18,7 +18,7 @@ public class Seed {
     private String seedContent;
 
     @Column(name = "seed_price")
-    private Integer seedPrice;
+    private Double seedPrice;
 
     @Column(name = "seed_period")
     private Long seedPeriod;
